requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.4
django-rest-knox==4.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
idna==2.8
mysqlclient==1.4.4
oauthlib==3.1.0
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.3
